+++ structures/Authenticator.js
@@ -1,9 +1,0 @@
-const authenticator = require('authenticator')
-
-class Authenticator {
-  static GetAuthToken (key) {
-    return authenticator.generateToken(key)
-  };
-};
-
-module.exports = Authenticator