+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # fnbot-client - Changelogs
 
+## Version 1.3
+
+- Using [fnbr.js](https://fnbr.js.org/) instead of `epicgames-client` and `epicgames-fortnite-client`. (New library that is up to date)
+- Commands can now be used in party chat
+- Fixed some issues
+- Login reworked. You'll now have to use a link one time to login. After that, device auth will be used to login.
+- Build information is no longer used. The cosmetics API is still used to search cosmetics.
+- Some minor changes to improve performance and quality
+
+> Reminder: Some issues could happen, the library is pretty new and the author is working on bug fixes already.
+
 ## Version 1.2.2
 
 - Added romanian translation. Thanks to `Skadro#1659`.